tests passing and we now fail the module when we try to delete tables or buckets without force

+++ src/index.js
@@ -2,7 +2,15 @@ let aws = require('aws-sdk')
 let waterfall = require('run-waterfall')
 let deleteBucketContents = require('./_delete-bucket-contents')
 
+/**
+ * @param {object} params - named parameters
+ * @param {string} params.name - name of cloudformation stack to delete
+ * @param {boolean} params.force - delete regardless of tables or buckets
+ */
 module.exports = function nuke(params, callback) {
+
+  // deletes buckets and tables with impunity
+  let force = params.force || false
 
   // always validate input
   if (!params.name)
@@ -43,26 +51,51 @@ module.exports = function nuke(params, callback) {
 
     // delete static assets
     function(bucketExists, callback) {
-      if (bucketExists) {
+      if (bucketExists && force) {
         let bucket = bucketExists.OutputValue.replace('http://', '').split('.')[0]
         deleteBucketContents({ 
           bucket
         }, callback)
       }
+      else if (bucketExists && force === false) {
+        // throw a big error here
+        callback(Error('found bucket, use --force to delete')) 
+      }
       else {
+        console.log('no bucket wtf?')
         callback()
       }
     },
 
-    // delete actual stack
     function(callback) {
-      cloudformation.deleteStack({
-        StackName: params.name, 
+      cloudformation.describeStackResources({
+        StackName: params.name 
       }, 
-      function(err) {
+      function(err, data) {
         if (err) callback(err)
-        else callback()
+        else {
+          let type = t=> t.ResourceType
+          let table = i=> i === 'AWS::DynamoDB::Table'
+          let hasTables = data.StackResources.map(type).some(table)
+          callback(null, hasTables)
+        }
       })
+    },
+
+    function(hasTables, callback) {
+      if (hasTables && force === false) {
+        callback(Error('found tables, use --force to delete'))
+      }
+      else {
+        // got this far, delete everything
+        cloudformation.deleteStack({
+          StackName: params.name, 
+        }, 
+        function(err) {
+          if (err) callback(err)
+          else callback()
+        })
+      }
     },
 
     // poll for progress
